Update ai requests resources
