+++ plugins/ai-bot-resources/package.json
@@ -40,11 +40,12 @@
   "dependencies": {
     "@hcengineering/ai-bot": "^0.6.0",
     "@hcengineering/analytics-collector": "^0.6.0",
+    "@hcengineering/chunter": "^0.6.20",
     "@hcengineering/core": "^0.6.32",
+    "@hcengineering/love": "^0.6.0",
     "@hcengineering/platform": "^0.6.11",
     "@hcengineering/presentation": "^0.6.3",
     "@hcengineering/ui": "^0.6.15",
-    "@hcengineering/chunter": "^0.6.20",
     "svelte": "^4.2.12"
   }
 }
