+++ plugins/ai-bot-resources/src/index.ts
@@ -16,7 +16,7 @@
 import { type Resources } from '@hcengineering/platform'
 import OnboardingChannelPanelExtension from './components/OnboardingChannelAsideExtension.svelte'
 
-export * from './utils'
+export * from './requests'
 
 export default async (): Promise<Resources> => ({
   component: {
